# routers/api/docs/swagger.yaml
basePath: /
definitions:
  Rsp:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/proto.CodeType'
        description: msg code
      message:
        description: msg info
        type: string
      requestId:
        description: trans id
        type: string
      results:
        description: response body
    type: object
  async.AsyncTask:
    properties:
      processed:
        type: integer
      taskId:
        type: string
      taskStatus:
        type: string
      total:
        type: integer
    type: object
...
  proto.CodeType:
    enum:
    - 200
    - 201
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 1009
    - 1010
    - 1011
    - 1012
    - 1013
    - 1014
    - 1015
    - 1016
    - 1017
    - 1018
    - 1021
    - 1022
    - 1027
    - 1028
    - 1029
    - 1030
    - 1031
    - 1032
    - 1036
    - 1037
    - 1038
    - 1046
    - 1050
    - 1061
    - 1062
    type: integer
    x-enum-comments:
      CodeCopyIdError: 文件夹不能拷贝到自己
      CodeFailedToConnect: 建立连接失败
      CodeFailedToConnectDB: 连接DB失败
      CodeFailedToCreateBucket: 失败去创建桶
      CodeFailedToCreateFolder: 失败去创建文件夹
      CodeFailedToCreateMultipartTask: 失败去创建任务
      CodeFailedToCreateSymlink: 失败去创建符号链接
      CodeFailedToDeleteUser: 删除用户失败
      CodeFailedToGetUsedStorage: 获取用户空间使用量失败
      CodeFailedToInitUser: 失败去初始化用户
      CodeFailedToOpenFile: 打开文件失败
      CodeFailedToOperateDB: 失败去操作数据库
      CodeFailedToOptMinio: 失败去写文件
      CodeFailedToSaveFile: 失败去保存数据
      CodeFileExist: 文件已存在
      CodeFileNotExist: 文件不存在
      CodeFolderDepthTooLong: 文件夹层数超过5层
      CodeFolderExist: 文件夹已存在
      CodeFolderNotExist: 文件夹不存在
      CodeGetAsyncTaskInfoFailed: 获取异步任务状态失败
      CodeMultipartRangeUploaded: 分片范围已上传
      CodeMultipartTaskCompleteErr: 合并错误
      CodeMultipartTaskHashErr: 文件hash校验错误
      CodeMultipartTaskNotFound: 未找到任务
      CodeMultipartTaskOverlap: 任务重叠
      CodeMultipartTaskRangeErr: 上传范围错误
      CodeMultipartUploadingConflit: 分片范围上传冲突
      CodeNotEnoughSpace: 空间不够，不上传
      CodeParamErr: 参数错误
      CodeReqParamErr: 请求参数错误(推荐)
      CodeUserIdError: 用户Id的错误
      CodeWriteRedisFailed: 写入redis 失败
      CodeZipFileFailed: 压缩文件失败
    x-enum-varnames:
    - CodeOk
    - CodeCreateAsyncTaskSuccess
    - CodeParamErr
    - CodeReqParamErr
    - CodeFileNotExist
    - CodeFailedToOpenFile
    - CodeFolderNotExist
    - CodeFailedToConnect
    - CodeFailedToConnectDB
    - CodeFailedToOptMinio
    - CodeFailedToCreateBucket
    - CodeFailedToSaveFile
    - CodeFailedToOperateDB
    - CodeFailedToCreateFolder
    - CodeFileExist
    - CodeFolderExist
    - CodeFolderDepthTooLong
    - CodeUserIdError
    - CodeFailedToInitUser
    - CodeFailedToDeleteUser
    - CodeFailedToGetUsedStorage
    - CodeCopyIdError
    - CodeFailedToCreateMultipartTask
    - CodeMultipartTaskNotFound
    - CodeMultipartTaskOverlap
    - CodeMultipartTaskRangeErr
    - CodeMultipartTaskHashErr
    - CodeMultipartTaskCompleteErr
    - CodeNotEnoughSpace
    - CodeMultipartRangeUploaded
    - CodeMultipartUploadingConflit
    - CodeWriteRedisFailed
    - CodeZipFileFailed
    - CodeFailedToCreateSymlink
    - CodeGetAsyncTaskInfoFailed
...
  proto.FileInfoPub:
    properties:
      betag:
        type: string
      category:
        type: string
      createdAt:
        type: integer
      fileCount:
        type: integer
      isDir:
        type: boolean
      mime:
        type: string
      modifyAt:
        type: integer
      name:
        type: string
      operationAt:
        type: integer
      parentUuid:
        type: string
      path:
        type: string
      size:
        type: integer
      trashed:
        description: 0-normal; 1-Logical delete, put into the recycle bin; 2-Has been
          cleared from the recycle bin and is to be physically deleted
        type: integer
      uuid:
        type: string
    type: object
...
  proto.GetListRspData:
    properties:
      fileList:
        items:
          $ref: '#/definitions/proto.FileInfoPub'
        type: array
      pageInfo:
        $ref: '#/definitions/proto.PageInfoExt'
    type: object
...
  proto.PageInfoExt:
    properties:
      count:
        type: integer
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  proto.RecycledPhyDeleteReq:
    properties:
      uuids:
        items:
          type: string
        type: array
    type: object
  proto.RestoreRecycledReq:
    properties:
      uuids:
        items:
          type: string
        type: array
    required:
    - uuids
    type: object
  proto.StatusRsp:
    properties:
      status:
        type: string
    type: object
  proto.Storage:
    properties:
      userStorage:
        type: integer
    type: object
  proto.User:
    properties:
      userId:
        type: integer
    type: object
...
info:
  contact:
    email: service@ao.space
    name: AO.space
    url: https://ao.space/
  description: This is AO.space aofs OpenAPI reference document.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: aofs apis
  version: "1.0"
paths:
...
  /space/v1/api/recycled/clear:
    post:
      consumes:
      - application/json
      description: Clean out the recycle bin
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: integer
      - description: params
        in: body
        name: uuids
        required: true
        schema:
          $ref: '#/definitions/proto.RecycledPhyDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Rsp'
      summary: Clean out the recycle bin
      tags:
      - Recycled
  /space/v1/api/recycled/list:
    get:
      consumes:
      - application/json
      description: Query the recycle bin file list
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      - description: 'page index，default: 1'
        in: query
        name: page
        type: integer
      - description: 'page size，default: 10'
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/proto.GetListRspData'
              type: object
      summary: Query the recycle bin file list
      tags:
      - Recycled
  /space/v1/api/recycled/restore:
    post:
      consumes:
      - application/json
      description: Restore files from recycle bin
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: integer
      - description: params
        in: body
        name: restoreFilesReq
        required: true
        schema:
          $ref: '#/definitions/proto.RestoreRecycledReq'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/Rsp'
        "201":
          description: asynchronous task
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/async.AsyncTask'
              type: object
      summary: Restore files from recycle bin
      tags:
      - Recycled
  /space/v1/api/status:
    get:
      consumes:
      - application/json
      description: Querying service status
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/proto.StatusRsp'
              type: object
      summary: Querying service status
      tags:
      - ServerStatus
  /space/v1/api/sync/synced:
    get:
      consumes:
      - application/json
      description: Obtain incremental synchronization data
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      - description: timestamp
        in: query
        name: timestamp
        required: true
        type: integer
      - description: path
        in: query
        name: path
        type: string
      - description: device id
        in: query
        name: deviceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/proto.GetListRspData'
              type: object
      summary: Obtain incremental synchronization data
      tags:
      - Sync
  /space/v1/api/user/delete:
    post:
      consumes:
      - application/json
      description: Delete user
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      - description: user id
        in: body
        name: userId
        required: true
        schema:
          $ref: '#/definitions/proto.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Rsp'
      summary: Delete user
      tags:
      - User
  /space/v1/api/user/init:
    post:
      consumes:
      - application/json
      description: Initialize user
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      - description: user id
        in: body
        name: userId
        required: true
        schema:
          $ref: '#/definitions/proto.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Rsp'
      summary: Initialize user
      tags:
      - User
  /space/v1/api/user/storage:
    get:
      consumes:
      - application/json
      description: Query the user space capacity
      parameters:
      - description: user id
        in: query
        name: userId
        required: true
        type: string
      - description: user id
        in: query
        name: targetUserId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/Rsp'
            - properties:
                results:
                  $ref: '#/definitions/proto.Storage'
              type: object
      summary: Query the user space capacity
      tags:
      - User
swagger: "2.0"